+++ packages/grid/src/GridItem.tsx
@@ -1,3 +1,4 @@
-import { Block } from '@nawt/block';
+import styled from 'styled-components';
+import { styledObject } from '@nawt/styled';
 
-export const GridItem = Block;
+export const GridItem = styled(styledObject('div', {}))``;