+++ packages/text/src/index.ts
@@ -1,1 +1,0 @@
-export { TextBase, Text } from './Text';