+++ packages/block/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nawt/block
 
+## 0.1.2
+
+### Patch Changes
+
+- Configuration Update
+- Block component fix
+
 ## 0.1.1
 
 ### Patch Changes