+++ packages/flex/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nawt/flex
 
+## 0.1.2
+
+### Patch Changes
+
+- Configuration Update
+
 ## 0.1.1
 
 ### Patch Changes