+++ packages/text/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nawt/text
 
+## 0.1.2
+
+### Patch Changes
+
+- Configuration Update
+- Remove TextBase export
+
 ## 0.1.1
 
 ### Patch Changes